refactor(workflow): keep the preview resets in the effect component
Closing the confirmation belongs with the preview reset rather than in
the component that renders the actions, so both live in the effect
component and the modal id moves to its own constant.
