+++ packages/twenty-front/src/modules/object-core/workflows/versions/constants/RestoreWorkflowVersionModalId.ts
@@ -1,0 +1,2 @@
+export const RESTORE_WORKFLOW_VERSION_MODAL_ID =
+  'restore-workflow-version-modal';
